requirements.txt
Flask
Authlib
Flask_SQLAlchemy
Flask-Session
Flask-Caching
Flask-Sockets
redis
smart_getenv
SQLAlchemy
Werkzeug
python-etcd
flask_mako
raven
PyYAML
more_itertools
gunicorn
pymysql
requests
celery[redis]
humanize
humanfriendly
boltons
python-box
blinker
wrapt
pyparsing
crontab
webargs
marshmallow
delegator.py
kubernetes